requirements.txt
PyQt5==5.15.11
python-dotenv==1.1.1